requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.4
psycopg2-binary==2.9.7
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-Babel==3.0.1
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
Flask-Uploads==0.2.1
marshmallow==3.19.0
WeasyPrint==57.2
pandas==2.2.3
openpyxl==3.1.2
Babel==2.12.1
bcrypt==4.0.1
